move gauth related decorators to gauth dir. refactor gauth_required for higher generality.
